"Merging to get updated version"

+++ src/ieee.h
@@ -1,6 +1,8 @@
 #ifndef IEEE_H
 #define IEEE_H
 
-bool feq(double a,double b);
+#include "misc.h"
+
+bool feq(float a,float b);
 
 #endif //IEEE_H
